# strava/aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Strava API v3",
    "description" : "The [Swagger Playground](https://developers.strava.com/playground) is the easiest way to familiarize yourself with the Strava API by submitting HTTP requests and observing the responses before you write any client code. It will show what a response will look like with different endpoints depending on the authorization scope you receive from your athletes. To use the Playground, go to https://www.strava.com/settings/api and change your “Authorization Callback Domain” to developers.strava.com. Please note, we only support Swagger 2.0. There is a known issue where you can only select one scope at a time. For more information, please check the section “client code” at https://developers.strava.com/docs.",
    "version" : "3.0.0"
  },
  "servers" : [ {
    "url" : "https://www.strava.com/api/v3"
  } ],
  "security" : [ {
    "strava_oauth" : [ "public" ]
  } ],
  "paths" : {
...
    "/uploads/{uploadId}" : {
      "get" : {
        "tags" : [ "Uploads" ],
        "summary" : "Get Upload",
        "description" : "Returns an upload for a given identifier. Requires activity:write scope.",
        "operationId" : "getUploadById",
        "parameters" : [ {
          "name" : "uploadId",
          "in" : "path",
          "description" : "The identifier of the upload.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Representation of the upload.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Upload"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/activities/{id}/streams" : {
      "get" : {
        "tags" : [ "Streams" ],
        "summary" : "Get Activity Streams",
        "description" : "Returns the given activity's streams. Requires activity:read scope. Requires activity:read_all scope for Only Me activities.",
        "operationId" : "getActivityStreams",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the activity.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "keys",
          "in" : "query",
          "description" : "Desired stream types.",
          "required" : true,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "time", "distance", "latlng", "altitude", "velocity_smooth", "heartrate", "cadence", "watts", "temp", "moving", "grade_smooth" ]
            }
          }
        }, {
          "name" : "key_by_type",
          "in" : "query",
          "description" : "Must be true.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The set of requested streams.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StreamSet"
                },
                "example" : [ {
                  "type" : "distance",
                  "data" : [ 2.9, 5.8, 8.5, 11.7, 15, 19, 23.2, 28, 32.8, 38.1, 43.8, 49.5 ],
                  "series_type" : "distance",
                  "original_size" : 12,
                  "resolution" : "high"
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/segment_efforts/{id}/streams" : {
      "get" : {
        "tags" : [ "Streams" ],
        "summary" : "Get Segment Effort Streams",
        "description" : "Returns a set of streams for a segment effort completed by the authenticated athlete. Requires read_all scope.",
        "operationId" : "getSegmentEffortStreams",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the segment effort.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "keys",
          "in" : "query",
          "description" : "The types of streams to return.",
          "required" : true,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "time", "distance", "latlng", "altitude", "velocity_smooth", "heartrate", "cadence", "watts", "temp", "moving", "grade_smooth" ]
            }
          }
        }, {
          "name" : "key_by_type",
          "in" : "query",
          "description" : "Must be true.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The set of requested streams.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StreamSet"
                },
                "example" : [ {
                  "type" : "distance",
                  "data" : [ 904.5, 957.8, 963.1, 989.1, 1011.9, 1049.7, 1082.4, 1098.1, 1113.2, 1124.7, 1139.2, 1142.1, 1170.4, 1173 ],
                  "series_type" : "distance",
                  "original_size" : 14,
                  "resolution" : "high"
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/segments/{id}/streams" : {
      "get" : {
        "tags" : [ "Streams" ],
        "summary" : "Get Segment Streams",
        "description" : "Returns the given segment's streams. Requires read_all scope for private segments.",
        "operationId" : "getSegmentStreams",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the segment.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "keys",
          "in" : "query",
          "description" : "The types of streams to return.",
          "required" : true,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "distance", "latlng", "altitude" ]
            }
          }
        }, {
          "name" : "key_by_type",
          "in" : "query",
          "description" : "Must be true.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The set of requested streams.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StreamSet"
                },
                "example" : [ {
                  "type" : "latlng",
                  "data" : [ [ 37.833112, -122.483436 ], [ 37.832964, -122.483406 ] ],
                  "series_type" : "distance",
                  "original_size" : 2,
                  "resolution" : "high"
                }, {
                  "type" : "distance",
                  "data" : [ 0, 16.8 ],
                  "series_type" : "distance",
                  "original_size" : 2,
                  "resolution" : "high"
                }, {
                  "type" : "altitude",
                  "data" : [ 92.4, 93.4 ],
                  "series_type" : "distance",
                  "original_size" : 2,
                  "resolution" : "high"
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/routes/{id}/streams" : {
      "get" : {
        "tags" : [ "Streams" ],
        "summary" : "Get Route Streams",
        "description" : "Returns the given route's streams. Requires read_all scope for private routes.",
        "operationId" : "getRouteStreams",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the route.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The set of requested streams.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StreamSet"
                },
                "example" : [ {
                  "type" : "latlng",
                  "data" : [ [ 37.833112, -122.483436 ], [ 37.832964, -122.483406 ] ]
                }, {
                  "type" : "distance",
                  "data" : [ 0, 16.8 ]
                }, {
                  "type" : "altitude",
                  "data" : [ 92.4, 93.4 ]
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Fault" : {
        "type" : "object",
        "properties" : {
          "errors" : {
            "type" : "array",
            "description" : "The set of specific errors associated with this fault, if any.",
            "items" : {
              "$ref" : "#/components/schemas/Error"
            }
          },
          "message" : {
            "type" : "string",
            "description" : "The message of the fault."
          }
        },
        "description" : "Encapsulates the errors that may be returned from the API."
      },
      "Error" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "The code associated with this error."
          },
          "field" : {
            "type" : "string",
            "description" : "The specific field or aspect of the resource associated with this error."
          },
          "resource" : {
            "type" : "string",
            "description" : "The type of resource associated with this error."
          }
        }
      },
...
      "LatLng" : {
        "maxItems" : 2,
        "minItems" : 2,
        "type" : "array",
        "description" : "A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.",
        "items" : {
          "type" : "number",
          "format" : "float"
        }
      },
...
      "Upload" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "The unique identifier of the upload",
            "format" : "int64"
          },
          "id_str" : {
            "type" : "string",
            "description" : "The unique identifier of the upload in string format"
          },
          "external_id" : {
            "type" : "string",
            "description" : "The external identifier of the upload"
          },
          "error" : {
            "type" : "string",
            "description" : "The error associated with this upload"
          },
          "status" : {
            "type" : "string",
            "description" : "The status of this upload"
          },
          "activity_id" : {
            "type" : "integer",
            "description" : "The identifier of the activity this upload resulted into",
            "format" : "int64"
          }
        },
        "example" : {
          "id_str" : "id_str",
          "activity_id" : 6,
          "external_id" : "external_id",
          "id" : 0,
          "error" : "error",
          "status" : "status"
        }
      },
      "StreamSet" : {
        "type" : "object",
        "properties" : {
          "time" : {
            "$ref" : "#/components/schemas/TimeStream"
          },
          "distance" : {
            "$ref" : "#/components/schemas/DistanceStream"
          },
          "latlng" : {
            "$ref" : "#/components/schemas/LatLngStream"
          },
          "altitude" : {
            "$ref" : "#/components/schemas/AltitudeStream"
          },
          "velocity_smooth" : {
            "$ref" : "#/components/schemas/SmoothVelocityStream"
          },
          "heartrate" : {
            "$ref" : "#/components/schemas/HeartrateStream"
          },
          "cadence" : {
            "$ref" : "#/components/schemas/CadenceStream"
          },
          "watts" : {
            "$ref" : "#/components/schemas/PowerStream"
          },
          "temp" : {
            "$ref" : "#/components/schemas/TemperatureStream"
          },
          "moving" : {
            "$ref" : "#/components/schemas/MovingStream"
          },
          "grade_smooth" : {
            "$ref" : "#/components/schemas/SmoothGradeStream"
          }
        },
        "example" : {
          "altitude" : "",
          "temp" : "",
          "velocity_smooth" : "",
          "watts" : "",
          "distance" : "",
          "heartrate" : "",
          "time" : "",
          "moving" : "",
          "latlng" : "",
          "cadence" : "",
          "grade_smooth" : ""
        }
      },
      "TimeStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of time values for this stream, in seconds",
              "items" : {
                "type" : "integer"
              }
            }
          }
        } ]
      },
      "BaseStream" : {
        "type" : "object",
        "properties" : {
          "original_size" : {
            "type" : "integer",
            "description" : "The number of data points in this stream"
          },
          "resolution" : {
            "type" : "string",
            "description" : "The level of detail (sampling) in which this stream was returned",
            "enum" : [ "low", "medium", "high" ]
          },
          "series_type" : {
            "type" : "string",
            "description" : "The base series used in the case the stream was downsampled",
            "enum" : [ "distance", "time" ]
          }
        }
      },
      "DistanceStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of distance values for this stream, in meters",
              "items" : {
                "type" : "number",
                "format" : "float"
              }
            }
          }
        } ]
      },
      "LatLngStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of lat/long values for this stream",
              "items" : {
                "$ref" : "#/components/schemas/LatLng"
              }
            }
          }
        } ]
      },
      "AltitudeStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of altitude values for this stream, in meters",
              "items" : {
                "type" : "number",
                "format" : "float"
              }
            }
          }
        } ]
      },
      "SmoothVelocityStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of velocity values for this stream, in meters per second",
              "items" : {
                "type" : "number",
                "format" : "float"
              }
            }
          }
        } ]
      },
      "HeartrateStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of heart rate values for this stream, in beats per minute",
              "items" : {
                "type" : "integer"
              }
            }
          }
        } ]
      },
      "CadenceStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of cadence values for this stream, in rotations per minute",
              "items" : {
                "type" : "integer"
              }
            }
          }
        } ]
      },
      "PowerStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of power values for this stream, in watts",
              "items" : {
                "type" : "integer"
              }
            }
          }
        } ]
      },
      "TemperatureStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of temperature values for this stream, in celsius degrees",
              "items" : {
                "type" : "integer"
              }
            }
          }
        } ]
      },
      "MovingStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of moving values for this stream, as boolean values",
              "items" : {
                "type" : "boolean"
              }
            }
          }
        } ]
      },
      "SmoothGradeStream" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/BaseStream"
        }, {
          "type" : "object",
          "properties" : {
            "data" : {
              "type" : "array",
              "description" : "The sequence of grade values for this stream, as percents of a grade",
              "items" : {
                "type" : "number",
                "format" : "float"
              }
            }
          }
        } ]
      },
...
    },
...
    "securitySchemes" : {
      "strava_oauth" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://www.strava.com/api/v3/oauth/authorize",
            "tokenUrl" : "https://www.strava.com/api/v3/oauth/token",
            "scopes" : {
              "read" : "Read public segments, public routes, public profile data, public posts, public events, club feeds, and leaderboards",
              "read_all" : "Read private routes, private segments, and private events for the user",
              "profile:read_all" : "Read all profile information even if the user has set their profile visibility to Followers or Only You",
              "profile:write" : "Update the user's weight and Functional Threshold Power (FTP), and access to star or unstar segments on their behalf",
              "activity:read" : "Read the user's activity data for activities that are visible to Everyone and Followers, excluding privacy zone data",
              "activity:read_all" : "The same access as activity:read, plus privacy zone data and access to read the user's activities with visibility set to Only You",
              "activity:write" : "Access to create manual activities and uploads, and access to edit any activities that are visible to the app, based on activity read access level"
            }
          }
        }
      }
    }
  },
...
}